EXPORT_PY/requirements.txt
Pillow==10.1.0
Pillow==10.2.0
python_tkdnd==0.2.1
tkinterDnD==0.0.0
